Move the `api` code away from `lib`
